John Cage Simulator: Creates a piece of variable length from randomly chosen clips of Cage's work.
For each run of the program, a random number of I-Ching hexagrams are chosen, which causes the length of the piece to vary for each run. Each of the hexagrams corresponds to an audio clip of Cage's music. The positionof the I-Ching card in the sequence corresponds to the place of the clip in the piece. During the run, a graphic displays the I-Ching hexagrams used to construct the piece and the audio clips are played in their determined order.